[stm32] actuators_pwm: replace set_servo_gpio with gpio_setup_pin_af
- AFIO remap in board file
- gpio_setup_pin_af with uint32_t for af
- remove PWM_SERVO_x_RCC

+++ sw/airborne/arch/stm32/mcu_periph/gpio_arch.c
@@ -86,7 +86,7 @@ void gpio_setup_input(uint32_t port, uint16_t gpios) {
   gpio_set_mode(port, GPIO_MODE_INPUT, GPIO_CNF_INPUT_FLOAT, gpios);
 }
 
-void gpio_setup_pin_af(uint32_t port, uint16_t pin, uint8_t af, bool_t is_output) {
+void gpio_setup_pin_af(uint32_t port, uint16_t pin, uint32_t af, bool_t is_output) {
   gpio_enable_clock(port);
   /* remap alternate function if needed */
   if (af) {
